Modernization
We are moving to new and better world!

This no longer runs in Docker; instead, it runs the backend in Lambda and the frontend as a static site in S3.

There are new build/deploy scripts for both the backend and frontend.

Also, the React dependency is updated, and we are using React.PureComponent rather than the older React.createClass. Instead of our homegrown dispatcher/stores solution, we are now using a cleaner and more modern Redux/reducer setup; the geolocation and network traffic is hooked in via middleware.